is_keuken gerefactored naar gedeelde_logica
