+++ woningwaardering/stelsels/gedeelde_logica/keuken/__init__.py
@@ -1,0 +1,3 @@
+from .keuken import is_keuken
+
+__all__ = ["is_keuken"]
